+++ src/relativePaths.ts
@@ -1,0 +1,44 @@
+import * as vscode from 'vscode';
+import * as utils from './utils';
+
+const getRelPaths = (files: Array<vscode.Uri>, workspace: vscode.WorkspaceFolder) => {
+  if(files.length){
+      
+      if(workspace){
+          const regex = new RegExp(workspace.uri.path,'g');
+
+          const relPaths =files.map((element: vscode.Uri) => element.path.replace(regex,''));
+          return relPaths;
+      }else{
+          vscode.window.showErrorMessage('Cant find workspace');
+      }
+  } else {
+      vscode.window.showErrorMessage('Could not copy');
+  }
+  return null;
+};
+
+const getFocusedRelativePaths = (relPaths: Array<string>, workspace: vscode.WorkspaceFolder) => {
+  const activeTextEditor = vscode.window.activeTextEditor;
+  if(activeTextEditor) {
+      const focusedFile = activeTextEditor.document.fileName.replace(workspace.uri.path,'');
+        return relPaths.map((path: string) => {
+          const regex = new RegExp(utils.sharedStart([path.replace(workspace.uri.path,''), focusedFile]).slice(0, -1),'g');
+          const focusedFileParts = utils.removeInPathAndSplit(focusedFile, regex);
+          let pathParts = utils.removeInPathAndSplit(path, regex);
+
+          if(focusedFileParts.length === 1){
+              return './' + pathParts.join('/');
+          } else {
+              return '../'.repeat(focusedFileParts.length-1) + pathParts.join('/');
+          }
+      });
+  }
+  return null;
+};
+
+export {
+  getRelPaths,
+  getFocusedRelativePaths
+
+};