+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [0.1.0]
+* Added: `Copy Relative On Focus` - gets relative file to your focused file
+* Added: basic automated testing
+
  ## 0.0.5
  * Added Keybind `ctrl+shift+r` or `cmd+shift+r`
  * Improved docs